// komg/komgo-member-api-coverage/src/business-layer/notifications/enums.ts
export enum PRODUCT_ID {
  Coverage = 'coverage'
}

export enum NOTIFICATION_TYPE {
  COUNTERPARTY_INFO = 'Counterparty.info',
  COUNTERPARTY_TASK = 'Counterparty.task'
}

export enum NOTIFICATION_USER {
  // CoverageCrud = 'manageCoverage.crud',
  // CoverageRead = 'manageCoverage.read',
  CoverageManage = 'manageCoverage'
}

export enum NOTIFICATION_LEVEL {
  // success = 'success',
  info = 'info',
  warning = 'warning',
  danger = 'danger'
}